Add files via upload
Initial IRB protocol was modified to address the unexpected sharing of IP addresses by Qualtrics. These data were collected for quality control to ensure that more than one survey was not submitted in a short period of time from the same computer. Our new protocol is to exclude these data from the downloaded data used for analysis. We will not be viewing these data and they will not be included in the shared data. This letter is the approval for the new protocol for handling the unexpected data.